requirements.txt
Mastodon.py==1.8.0
git+https://github.com/XaviArnaus/pyxavi@v0.1.3